organized, added tests
