+++ contrib/render/test/test_scene.py
@@ -1,0 +1,11 @@
+import numpy as np
+from vispy import scene, app
+from vispy.geometry import MeshData
+from vispy.scene import visuals
+from contrib.render.render.scene import MainWindow
+
+# Dont know how to write scene test
+def test_set_model():
+    pass
+
+test_set_model()
